feat: `Toolkit\HasI18n` trait
