+++ src/Panel/Ui/Component.php
@@ -3,7 +3,7 @@
 namespace Kirby\Panel\Ui;
 
 use Kirby\Exception\LogicException;
-use Kirby\Toolkit\I18n;
+use Kirby\Toolkit\HasI18n;
 use Kirby\Toolkit\Str;
 
 /**
@@ -19,6 +19,8 @@ use Kirby\Toolkit\Str;
  */
 abstract class Component
 {
+	use HasI18n;
+
 	protected string $key;
 	public array $attrs = [];
 
@@ -54,15 +56,6 @@ abstract class Component
 		// setter
 		$this->$name = $args[0];
 		return $this;
-	}
-
-	/**
-	 * Translates a key to a string
-	 * @since 6.0.0
-	 */
-	protected function i18n(string|array|null $key): string|null
-	{
-		return $key !== null ? I18n::translate($key, $key) : null;
 	}
 
 	/**
